+++ src/Api/UnsupportedOperationException.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace Puli\Repository;
+namespace Puli\Repository\Api;
 
 use RuntimeException;