+++ CHANGELOG.md
@@ -59,6 +59,8 @@ Changelog
  * merged `GenericResource` and `AbstractResource`
  * removed backend repositories from `InMemoryRepository` and `FilesystemRepository`
  * added symlink support to `FilesystemRepository`
+ * moved `UnsupportedOperationException` to `Api` namespace
+ * replaced `NoDirectoryException` by `UnsupportedOperationException`
 
 * 1.0.0-alpha4 (2014-12-03)